// package/src/types.ts
export interface AgentStatus {
  /** Agent ID */
  id: string;
  /** Brief description of current task */
  task: string;
  /** List of files currently being worked on */
  files: string[];
  /** List of files claimed by this agent */
  claimedFiles: string[];
  /** Channels this agent is listening to */
  channels: string[];
  /** Session start time */
  startedAt: string;
  /** Last heartbeat timestamp */
  lastHeartbeat: string;
}